[kernel] E: Update Return Type of signal_cond()
This commit modifies the signal_cond() kernel routine to return a 32-bit
value instead of a usize. This change enables more consistent handling
of return values across architectures and improves clarity in the
kernel's signaling semantics.

+++ src/kernel/src/pm/kcall/signal_cond.rs
@@ -22,12 +22,37 @@ use ::sys::{
 // Standalone Functions
 //==================================================================================================
 
+///
+/// # Description
+///
+/// Signals a condition variable.
+///
+/// # Parameters
+///
+/// - `pid`: Target process identifier.
+/// - `tid`: Target thread identifier.
+/// - `cond_addr`: Address of the condition variable.
+/// - `broadcast`: If `true`, wakes all waiting threads. Otherwise, wakes a single waiting thread.
+///
+/// # Returns
+///
+/// Upon successful completion, this function returns the number of threads that were awakened.
+/// Otherwise, it returns an error object that specifying the reason of failure.
+///
+/// # Safety
+///
+/// This function is unsafe because it operates on global variables.
+///
+/// It is safe to call this function if and only if the following conditions are met:
+///
+/// - The calling process does not hold a reference to the process manager.
+///
 pub unsafe fn signal_cond(
     pid: ProcessIdentifier,
     tid: ThreadIdentifier,
     cond_addr: usize,
     broadcast: bool,
-) -> Result<usize, Error> {
+) -> Result<u32, Error> {
     trace!(
         "signal_cond(): pid={pid:?}, tid={tid:?}, cond_addr={cond_addr:x?}, broadcast={broadcast}"
     );
@@ -35,7 +60,7 @@ pub unsafe fn signal_cond(
     // Unpack kernel call arguments.
     let cond_addr: ConditionAddress = ConditionAddress::from(cond_addr);
 
-    let awakened: usize = {
+    let awakened: u32 = {
         let cond: Condvar = ProcessManager::get_cond(cond_addr)?;
         if broadcast {
             cond.notify_all()?
